Use the onboard BRAM and USB serial on the upduino v2.
Change the serial pins to connect to the FTDI port, which are shared
with the SPI flash chip for programming the Upduino v2.

The SPI flash CSN pin must be driven high to release the pins, allowing
the ice40 to use them instead.

Remap the RGB LED to the leds[2:0] outputs.